fix(AVLTree): wrong order in ZRANGE output
Incorrect Comparison: In the insert method, the code was always comparing
the new node against the `root` of the tree instead of the current node during traversal.
This caused nodes to be inserted in the wrong positions, breaking the sorted order.

Typo in Balancing Logic: A typo in the rebalancing code `(getHight(unbalanced->left) - getHight(unbalanced->left))`
meant that the tree was never being correctly balanced after an insertion.